Move highstateconf_test tests into state_test.py
This keeps the tests organized and following the same naming convention
established elsewhere.
